Refactored and updating a customer's default source with testcard (adding a new token)
