+++ README.md
@@ -17,38 +17,25 @@ Commands are run at the command line to operate this system. For example:
 ## Commands
 
 ### Listing
-- `stripe_list_invoices.go`
-- _list of invoices and their details including paid status_
-- `stripe_list_charges.go`
-- _list all charges - charge ids can be used for refunds_
+- ls: list invoices
+- charges: list all charges for customer - charge ids can be used for refunds
+- getcustomer: lookup customer and show payment source details
 
 ### Invoicing
-- `stripe_invoice.go`
-  - _create a single invoice_
-- `stripe_invoice_offers.go`
-  - _create invoices from an offers csv file (see example csv)_
-- `stripe_finalize_draft_invoices.go`
-  - _finalize invoices to trigger email for payment_
-
-### Refunds
-- `stripe_refund.go`
-  - _refund a previous charge (payment taken by card)_
+- invoice: create a single invoice
+- offers: create invoices from an offers csv file (see example csv in ./test)
+- finalize: finalize invoices to trigger email for payment
+- void: void an invoice
 
 ### Testing
-- `stripe_test_addcard.go`
-  - _add a tokenized card for testing payments and refunds_
-- `stripe_test_invoice_pay.go`
-  - _test paying an invoice with a card previously added_
-- `offers-202001119.csv`
-  - _example offers file used for invoicing offers_
+- `offers-202001119.csv`: example offers file used for invoicing offers
 
 ### Utility functions
-- `./stripey/stripey.go`
-  - _common utility functions for stripe_
+- `./stripey/stripey.go`: common utility functions for stripe
 
 ## Setup Configuration
 - .env file containing the test and live keys
-- In the go.mod file add the following replacement:
+- In the go.mod file add the following replacement if the packages are not found:
 
 ```
 replace github.com/griffithbarracks/utils/stripey v0.0.1 => ./stripey
