+++ test/offers-20201119.csv
@@ -1,0 +1,3 @@
+S#,Childid,Child Name,Parent Name,Contact #,Email,Class,Seq,Draw,Activity,Amount,Donation,Invoice#,Invoice Status,Amount
+331,patricia.scanlon@gmail.com/laoisetay,Laoise Tay,Patricia Scanlon,0872659770,patricia.scanlon@gmail.com,4th Eoin,1,16,Chess : 3-6 Class,50.00,0.00,,#N/A,
+395,aaron.tay.ym@gmail.com/neoscanlon,Neo Scanlon,Aaron Tay,0872276226,aaron.tay.ym@gmail.com,1st Anne-Marie,16,11,Soccer : 1-3 Class,68.00,0.00,,#N/A,
